replace int with size_t

+++ test/test-text.cpp
@@ -10,55 +10,55 @@ using namespace algocpp::text;
 BOOST_AUTO_TEST_CASE(test_makeP) {
   {
     std::vector<size_t> expected_p {0};
-    std::vector<int> p(kmp::makeP(""));
+    std::vector<size_t> p(kmp::makeP(""));
     BOOST_CHECK_EQUAL_COLLECTIONS(p.begin(), p.end(), expected_p.begin(), expected_p.end());
   }
 
   {
     std::vector<size_t> expected_p {0,0};
-    std::vector<int> p(kmp::makeP("a"));
+    std::vector<size_t> p(kmp::makeP("a"));
     BOOST_CHECK_EQUAL_COLLECTIONS(p.begin(), p.end(), expected_p.begin(), expected_p.end());
   }
 
   {
     std::vector<size_t> expected_p {0,0,0};
-    std::vector<int> p(kmp::makeP("ab"));
+    std::vector<size_t> p(kmp::makeP("ab"));
     BOOST_CHECK_EQUAL_COLLECTIONS(p.begin(), p.end(), expected_p.begin(), expected_p.end());
   }
 
   {
     std::vector<size_t> expected_p {0,0,1};
-    std::vector<int> p(kmp::makeP("aa"));
+    std::vector<size_t> p(kmp::makeP("aa"));
     BOOST_CHECK_EQUAL_COLLECTIONS(p.begin(), p.end(), expected_p.begin(), expected_p.end());
   }
 
   {
     std::vector<size_t> expected_p {0,0,1,0};
-    std::vector<int> p(kmp::makeP("aab"));
+    std::vector<size_t> p(kmp::makeP("aab"));
     BOOST_CHECK_EQUAL_COLLECTIONS(p.begin(), p.end(), expected_p.begin(), expected_p.end());
   }
 
   {
     std::vector<size_t> expected_p {0,0,0,0};
-    std::vector<int> p(kmp::makeP("abb"));
+    std::vector<size_t> p(kmp::makeP("abb"));
     BOOST_CHECK_EQUAL_COLLECTIONS(p.begin(), p.end(), expected_p.begin(), expected_p.end());
   }
 
   {
     std::vector<size_t> expected_p {0,0,0,1};
-    std::vector<int> p(kmp::makeP("aba"));
+    std::vector<size_t> p(kmp::makeP("aba"));
     BOOST_CHECK_EQUAL_COLLECTIONS(p.begin(), p.end(), expected_p.begin(), expected_p.end());
   }
 
   {
     std::vector<size_t> expected_p {0,0,0,1,2,3};
-    std::vector<int> p(kmp::makeP("ababa"));
+    std::vector<size_t> p(kmp::makeP("ababa"));
     BOOST_CHECK_EQUAL_COLLECTIONS(p.begin(), p.end(), expected_p.begin(), expected_p.end());
   }
 
   {
     std::vector<size_t> expected_p {0,0,1,2,3,4};
-    std::vector<int> p(kmp::makeP("aaaaa"));
+    std::vector<size_t> p(kmp::makeP("aaaaa"));
     BOOST_CHECK_EQUAL_COLLECTIONS(p.begin(), p.end(), expected_p.begin(), expected_p.end());
   }
 }
